Clean up package and class structure significantly 1. Add service layer for backup / restore operations. 2. Reduce amount of duplicate code in activity classes.
