+++ app/src/main/java/com/stevesoltys/backup/activity/backup/BackupPopupWindowListener.java
@@ -4,7 +4,6 @@ import android.os.RemoteException;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-
 import com.stevesoltys.backup.R;
 import com.stevesoltys.backup.session.backup.BackupResult;
 import com.stevesoltys.backup.session.backup.BackupSession;
@@ -12,7 +11,7 @@ import com.stevesoltys.backup.session.backup.BackupSession;
 /**
  * @author Steve Soltys
  */
-class BackupPopupWindowListener implements Button.OnClickListener {
+public class BackupPopupWindowListener implements Button.OnClickListener {
 
     private static final String TAG = BackupPopupWindowListener.class.getName();
 
@@ -26,16 +25,13 @@ class BackupPopupWindowListener implements Button.OnClickListener {
     public void onClick(View view) {
         int viewId = view.getId();
 
-        switch (viewId) {
+        if (viewId == R.id.popup_cancel_button) {
+            try {
+                backupSession.stop(BackupResult.CANCELLED);
 
-            case R.id.popup_cancel_button:
-                try {
-                    backupSession.stop(BackupResult.CANCELLED);
-
-                } catch (RemoteException e) {
-                    Log.e(TAG, "Error cancelling backup session: ", e);
-                }
-                break;
+            } catch (RemoteException e) {
+                Log.e(TAG, "Error cancelling backup session: ", e);
+            }
         }
     }
 }
