+++ app/src/main/java/com/stevesoltys/backup/service/TransportService.java
@@ -1,35 +1,46 @@
-package com.stevesoltys.backup.session;
+package com.stevesoltys.backup.service;
 
 import android.app.backup.IBackupManager;
-import android.content.pm.IPackageManager;
-import android.content.pm.PackageInfo;
 import android.os.RemoteException;
 import android.os.ServiceManager;
 import com.stevesoltys.backup.session.backup.BackupSession;
 import com.stevesoltys.backup.session.backup.BackupSessionObserver;
 import com.stevesoltys.backup.session.restore.RestoreSession;
 import com.stevesoltys.backup.session.restore.RestoreSessionObserver;
+import com.stevesoltys.backup.transport.ConfigurableBackupTransport;
+import com.stevesoltys.backup.transport.ConfigurableBackupTransportService;
+import com.stevesoltys.backup.transport.component.BackupComponent;
+import com.stevesoltys.backup.transport.component.RestoreComponent;
+import com.stevesoltys.backup.transport.component.provider.ContentProviderBackupComponent;
+import com.stevesoltys.backup.transport.component.provider.ContentProviderBackupConfiguration;
+import com.stevesoltys.backup.transport.component.provider.ContentProviderRestoreComponent;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Set;
-
-import static android.os.UserHandle.USER_SYSTEM;
 
 /**
  * @author Steve Soltys
  */
-public class BackupManagerController {
+public class TransportService {
 
     private static final String BACKUP_TRANSPORT = "com.stevesoltys.backup.transport.ConfigurableBackupTransport";
 
     private final IBackupManager backupManager;
 
-    private final IPackageManager packageManager;
-
-    public BackupManagerController() {
+    public TransportService() {
         backupManager = IBackupManager.Stub.asInterface(ServiceManager.getService("backup"));
-        packageManager = IPackageManager.Stub.asInterface(ServiceManager.getService("package"));
+    }
+
+    public boolean initializeBackupTransport(ContentProviderBackupConfiguration configuration) {
+        ConfigurableBackupTransport backupTransport = ConfigurableBackupTransportService.getBackupTransport();
+
+        if (backupTransport.isActive()) {
+            return false;
+        }
+
+        BackupComponent backupComponent = new ContentProviderBackupComponent(configuration);
+        RestoreComponent restoreComponent = new ContentProviderRestoreComponent(configuration);
+        backupTransport.initialize(backupComponent, restoreComponent);
+        return true;
     }
 
     public BackupSession backup(BackupSessionObserver observer, Set<String> packages) throws RemoteException {
@@ -60,21 +71,5 @@ public class BackupManagerController {
         RestoreSession restoreSession = new RestoreSession(backupManager, observer, packages);
         restoreSession.start();
         return restoreSession;
-    }
-
-    public List<String> getEligiblePackages() throws RemoteException {
-        List<String> results = new ArrayList<>();
-        List<PackageInfo> packages = packageManager.getInstalledPackages(0, USER_SYSTEM).getList();
-
-        if (packages != null) {
-            for (PackageInfo packageInfo : packages) {
-
-                if (backupManager.isAppEligibleForBackup(packageInfo.packageName)) {
-                    results.add(packageInfo.packageName);
-                }
-            }
-        }
-
-        return results;
     }
 }
